Move more to eb extensions
