+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Services/SyntaxFacts/AbstractSyntaxFacts.cs
@@ -47,101 +47,7 @@ namespace Microsoft.CodeAnalysis.LanguageServices
 
         public abstract bool IsPreprocessorDirective(SyntaxTrivia trivia);
 
-        public bool ContainsInterleavedDirective(
-            ImmutableArray<SyntaxNode> nodes, CancellationToken cancellationToken)
-        {
-            if (nodes.Length > 0)
-            {
-                var span = TextSpan.FromBounds(nodes.First().Span.Start, nodes.Last().Span.End);
-
-                foreach (var node in nodes)
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-
-                    if (ContainsInterleavedDirective(span, node, cancellationToken))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        public bool ContainsInterleavedDirective(SyntaxNode node, CancellationToken cancellationToken)
-            => ContainsInterleavedDirective(node.Span, node, cancellationToken);
-
-        public bool ContainsInterleavedDirective(
-            TextSpan span, SyntaxNode node, CancellationToken cancellationToken)
-        {
-            foreach (var token in node.DescendantTokens())
-            {
-                if (ContainsInterleavedDirective(span, token, cancellationToken))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        protected abstract bool ContainsInterleavedDirective(TextSpan span, SyntaxToken token, CancellationToken cancellationToken);
-
-        public bool SpansPreprocessorDirective(IEnumerable<SyntaxNode> nodes)
-        {
-            if (nodes == null || nodes.IsEmpty())
-            {
-                return false;
-            }
-
-            return SpansPreprocessorDirective(nodes.SelectMany(n => n.DescendantTokens()));
-        }
-
-        /// <summary>
-        /// Determines if there is preprocessor trivia *between* any of the <paramref name="tokens"/>
-        /// provided.  The <paramref name="tokens"/> will be deduped and then ordered by position.
-        /// Specifically, the first token will not have it's leading trivia checked, and the last
-        /// token will not have it's trailing trivia checked.  All other trivia will be checked to
-        /// see if it contains a preprocessor directive.
-        /// </summary>
-        public bool SpansPreprocessorDirective(IEnumerable<SyntaxToken> tokens)
-        {
-            // we want to check all leading trivia of all tokens (except the 
-            // first one), and all trailing trivia of all tokens (except the
-            // last one).
-
-            var first = true;
-            var previousToken = default(SyntaxToken);
-
-            // Allow duplicate nodes/tokens to be passed in.  Also, allow the nodes/tokens
-            // to not be in any particular order when passed in.
-            var orderedTokens = tokens.Distinct().OrderBy(t => t.SpanStart);
-
-            foreach (var token in orderedTokens)
-            {
-                if (first)
-                {
-                    first = false;
-                }
-                else
-                {
-                    // check the leading trivia of this token, and the trailing trivia
-                    // of the previous token.
-                    if (SpansPreprocessorDirective(token.LeadingTrivia) ||
-                        SpansPreprocessorDirective(previousToken.TrailingTrivia))
-                    {
-                        return true;
-                    }
-                }
-
-                previousToken = token;
-            }
-
-            return false;
-        }
-
-        private bool SpansPreprocessorDirective(SyntaxTriviaList list)
-            => list.Any(t => IsPreprocessorDirective(t));
+        public abstract bool ContainsInterleavedDirective(TextSpan span, SyntaxToken token, CancellationToken cancellationToken);
 
         public abstract SyntaxList<SyntaxNode> GetAttributeLists(SyntaxNode node);
